+++ test/tst_dict_mem5.f90
@@ -33,7 +33,7 @@ contains
 
   subroutine mem(dealloc)
     logical, intent(in) :: dealloc
-    real(dp) :: va(400,400) ! roughly 1.22 MB
+    real(real64) :: va(400,400) ! roughly 1.22 MB
     type(dictionary_t) :: d1, d2
     va = 0.
     d1 = 'hello'.kv.va
